Refactor schema to use shared configuration and eliminate duplication
- Extract schema definition to shared src/lib/schema.ts file
- Update main app to import schema from shared configuration
- Update seed script to read schema/app configuration from shared file
- Eliminate schema duplication anti-pattern between app and seed script
- Maintain single source of truth for InstantDB schema and app ID
