+++ scripts/seed.js
@@ -1,9 +1,28 @@
 #!/usr/bin/env node
 
-import { init, i, id } from '@instantdb/admin'
+import { init, i } from '@instantdb/admin'
 import { randomUUID } from 'crypto'
+import { readFileSync } from 'fs'
+import { fileURLToPath } from 'url'
+import { dirname, join } from 'path'
 
-// Database schema definition for InstantDB
+// Get the directory of this script
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
+
+// Read the shared schema configuration
+const schemaConfigPath = join(__dirname, '../src/lib/schema.ts')
+const schemaConfigFile = readFileSync(schemaConfigPath, 'utf8')
+
+// Extract APP_ID from the shared config
+const appIdMatch = schemaConfigFile.match(/export const APP_ID = ['"`]([^'"`]+)['"`]/)
+if (!appIdMatch) {
+  throw new Error('Could not extract APP_ID from schema.ts')
+}
+const APP_ID = appIdMatch[1]
+
+// Recreate the schema using the admin SDK
+// This uses the same structure as defined in src/lib/schema.ts
 const _schema = i.schema({
   entities: {
     poses: i.entity({
@@ -33,9 +52,6 @@ const _schema = i.schema({
     })
   }
 })
-
-// Real InstantDB app ID
-const APP_ID = '63c65c15-20c2-418f-b504-a823ecadb2d0'
 
 const db = init({
   appId: APP_ID,
